+++ modal-go/image.go
@@ -71,7 +71,8 @@ func NewImageFromId(imageId string) *Image {
 	return &Image{ImageId: imageId}
 }
 
-func (image *Image) build(app *App) (*Image, error) {
+// Build eagerly builds an Image on Modal.
+func (image *Image) Build(app *App) (*Image, error) {
 	if image == nil {
 		return nil, InvalidError{"image must be non-nil"}
 	}